refactor: #60 remove logic getDataTable in service
